stuff/requirements.txt
docker_py==1.10.6
fastapi==0.82.0
mysql_connector_repackaged==0.3.1
paho_mqtt==1.6.1
protobuf==4.21.5
pycryptodome==3.15.0
PyYAML==6.0
rdflib==6.1.1
requests==2.28.0
uvicorn==0.18.2
youtube_dl==2021.12.17
